S5_01_Flask_CS_GOAT/requirements.txt
Flask>=2.2.3
Flask-SQLAlchemy>=3.0.0
python-dotenv>=1.0.0
prophet>=1.2.1
psycopg2-binary>=2.9.5